test: move Eloquent model test into its own class
Relocates the DB-backed model export/import test into a dedicated
EloquentModelTest so it doesn't raise FastExcelTest's object coupling
above the threshold (Codacy CouplingBetweenObjects). Cleaner separation
of the SQLite/Eloquent setup from the plain FastExcel tests.
